Using Lodash templates and css :afters
Now instead of adding strings together, we're using a lot more
reasonable _.template()s. These are probably a lot slower, but they're
not being used in any performance-critical areas of code.

We also now moved to having all style information in one place, as
before, half the style information was put in updateStyles() and the
rest was in renderGrid(). However, it's now all in updateStyles() thanks
to :after selectors.